backend/inventory/api/views/category.py
"""
ViewSet for Category model.
"""
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated
from ...domain.models import Category
from ..serializers.category import CategorySerializer


class CategoryViewSet(viewsets.ModelViewSet):
    """
    ViewSet for viewing and editing categories.
    """
    queryset = Category.objects.all()
    serializer_class = CategorySerializer
    permission_classes = [IsAuthenticated]
    filterset_fields = ['name']
    search_fields = ['name', 'description']
    ordering_fields = ['name', 'created_at']
